Re-jiggered keyboard into its own component
